+++ weblate/trans/tests/test_discovery.py
@@ -121,10 +121,18 @@ class ComponentDiscoveryTest(RepoTestCase):
         self.assertEqual(len(matched), 3)
         self.assertEqual(len(deleted), 0)
 
-        # Second discovery with restricted component match
+        # Remove some files
+        repository = self.component.repository
+        with repository.lock:
+            repository.remove(
+                ['po-link', 'second-po/cs.po', 'second-po/de.po'],
+                'Remove some files'
+            )
+
+        # Create new discover as it caches matches
         discovery = ComponentDiscovery(
             self.component,
-            r'(?P<component>po)/(?P<language>[^/]*)\.po',
+            r'(?P<component>[^/]*)/(?P<language>[^/]*)\.po',
             '{{ component|title }}',
             '^(?!xx).*$',
         )
@@ -134,19 +142,19 @@ class ComponentDiscoveryTest(RepoTestCase):
             preview=True, remove=True
         )
         self.assertEqual(len(created), 0)
-        self.assertEqual(len(matched), 0)
-        self.assertEqual(len(deleted), 3)
+        self.assertEqual(len(matched), 1)
+        self.assertEqual(len(deleted), 2)
 
         # Test component removal
         created, matched, deleted = discovery.perform(remove=True)
         self.assertEqual(len(created), 0)
-        self.assertEqual(len(matched), 0)
-        self.assertEqual(len(deleted), 3)
+        self.assertEqual(len(matched), 1)
+        self.assertEqual(len(deleted), 2)
 
         # Components should be now removed
         created, matched, deleted = discovery.perform(remove=True)
         self.assertEqual(len(created), 0)
-        self.assertEqual(len(matched), 0)
+        self.assertEqual(len(matched), 1)
         self.assertEqual(len(deleted), 0)
 
     def test_duplicates(self):