+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2018.
 * Add install_addon management command for automated addon installation.
 * Allow more fine grained ratelimit settings.
 * Added support for export and import of Excel files.
+* Improve component cleanup in case of multiple component discovery addons.
 
 weblate 3.1.1
 -------------